+++ lightningd/log.c
@@ -4,7 +4,6 @@
 #include <ccan/array_size/array_size.h>
 #include <ccan/err/err.h>
 #include <ccan/io/io.h>
-#include <ccan/list/list.h>
 #include <ccan/opt/opt.h>
 #include <ccan/read_write_all/read_write_all.h>
 #include <ccan/str/hex/hex.h>
@@ -19,6 +18,7 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <inttypes.h>
+#include <lightningd/json.h>
 #include <lightningd/jsonrpc.h>
 #include <lightningd/lightningd.h>
 #include <lightningd/options.h>
@@ -30,39 +30,6 @@
 
 /* Once we're up and running, this is set up. */
 struct log *crashlog;
-
-struct log_entry {
-	struct list_node list;
-	struct timeabs time;
-	enum log_level level;
-	unsigned int skipped;
-	const char *prefix;
-	char *log;
-	/* Iff LOG_IO */
-	const u8 *io;
-};
-
-struct log_book {
-	size_t mem_used;
-	size_t max_mem;
-	void (*print)(const char *prefix,
-		      enum log_level level,
-		      bool continued,
-		      const struct timeabs *time,
-		      const char *str,
-		      const u8 *io, size_t io_len,
-		      void *arg);
-	void *print_arg;
-	enum log_level print_level;
-	struct timeabs init_time;
-
-	struct list_head log;
-};
-
-struct log {
-	struct log_book *lr;
-	const char *prefix;
-};
 
 static void log_to_file(const char *prefix,
 			enum log_level level,
@@ -662,17 +629,6 @@ static void add_skipped(struct log_info *info)
 		json_object_end(info->response);
 		info->num_skipped = 0;
 	}
-}
-
-static void json_add_time(struct json_stream *result, const char *fieldname,
-			  struct timespec ts)
-{
-	char timebuf[100];
-
-	snprintf(timebuf, sizeof(timebuf), "%lu.%09u",
-		(unsigned long)ts.tv_sec,
-		(unsigned)ts.tv_nsec);
-	json_add_string(result, fieldname, timebuf);
 }
 
 static void log_to_json(unsigned int skipped,
