+++ lightningd/log.h
@@ -1,6 +1,7 @@
 #ifndef LIGHTNING_LIGHTNINGD_LOG_H
 #define LIGHTNING_LIGHTNINGD_LOG_H
 #include "config.h"
+#include <ccan/list/list.h>
 #include <ccan/tal/tal.h>
 #include <ccan/time/time.h>
 #include <ccan/typesafe_cb/typesafe_cb.h>
@@ -13,6 +14,39 @@ struct command;
 struct json_stream;
 struct lightningd;
 struct timerel;
+
+struct log_entry {
+	struct list_node list;
+	struct timeabs time;
+	enum log_level level;
+	unsigned int skipped;
+	const char *prefix;
+	char *log;
+	/* Iff LOG_IO */
+	const u8 *io;
+};
+
+struct log_book {
+	size_t mem_used;
+	size_t max_mem;
+	void (*print)(const char *prefix,
+		      enum log_level level,
+		      bool continued,
+		      const struct timeabs *time,
+		      const char *str,
+		      const u8 *io, size_t io_len,
+		      void *arg);
+	void *print_arg;
+	enum log_level print_level;
+	struct timeabs init_time;
+
+	struct list_head log;
+};
+
+struct log {
+	struct log_book *lr;
+	const char *prefix;
+};
 
 /* We can have a single log book, with multiple logs in it: it's freed by
  * the last struct log itself. */
