+++ lightningd/notification.h
@@ -3,6 +3,7 @@
 #include "config.h"
 #include <lightningd/jsonrpc.h>
 #include <lightningd/lightningd.h>
+#include <lightningd/log.h>
 #include <lightningd/plugin.h>
 
 bool notifications_have_topic(const char *topic);
@@ -10,5 +11,7 @@ bool notifications_have_topic(const char *topic);
 void notify_connect(struct lightningd *ld, struct node_id *nodeid,
 		    struct wireaddr_internal *addr);
 void notify_disconnect(struct lightningd *ld, struct node_id *nodeid);
+
+void notify_warning(struct lightningd *ld, struct log_entry *l);
 
 #endif /* LIGHTNING_LIGHTNINGD_NOTIFICATION_H */
